+++ src/Components/EmployeesFlow/Home/home.module.css
@@ -1,0 +1,69 @@
+.table {
+  width: 100%;
+  margin: 0 auto;
+}
+.buttonProfile {
+  width: 80px;
+  margin: 20px;
+  display: flex;
+  justify-content: flex-end;
+  align-items: flex-end;
+}
+.divEditProfile {
+  width: 100%;
+  display: flex;
+  justify-content: flex-end;
+}
+/* .buttonProfile {
+  width: 15%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 15px 10%;
+  height: 30px;
+  background-color: grey;
+  color: #FFF;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+} */
+
+.buttonProfile {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 15%;
+    height: 30px;
+    margin: 15px 10%;
+    background: rgb(171, 170, 170);
+    color: #fff !important;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    position: relative;
+    z-index: 1;
+    overflow: hidden;
+    display: inline-block;
+  }
+  .buttonProfile::after {
+    content: '';
+    background: #181823;
+    position: absolute;
+    z-index: -1;
+    padding: 16px 20px;
+    display: block;
+    top: 0;
+    bottom: 0;
+    left: -100%;
+    right: 100%;
+    -webkit-transition: all 0.35s;
+    transition: all 0.35s;
+  }
+  .buttonProfile:hover::after {
+    left: 0;
+    right: 0;
+    top: 0;
+    bottom: 0;
+    -webkit-transition: all 0.35s;
+    transition: all 0.35s;
+  }
