+++ src/Components/EmployeesFlow/Profile/profile.module.css
@@ -1,0 +1,162 @@
+h1 {
+  text-align: center;
+}
+.form {
+  border: 1px solid #ceced3;
+  border-radius: 4px;
+  box-shadow: inherit black;
+  display: flex;
+  flex-direction: column;
+  width: 80%;
+  align-self: center;
+  margin: 0 !important;
+}
+.label {
+  width: 45%;
+  padding-left: 60px;
+  padding-right: 10px;
+  font-family: 'Roboto';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 15px;
+  line-height: 100%;
+  color: #373867;
+}
+
+.row {
+  padding: 5px 0 10px 0;
+  height: auto;
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: center;
+  justify-content: space-between;
+  border-bottom: 1px solid #ceced3;
+}
+.text {
+  width: 45%;
+  text-align: left;
+  margin-block-start: 0px;
+  margin-block-end: 0px;
+  height: auto;
+}
+
+.form > :nth-child(even) {
+  background: #ffffff;
+  box-shadow: inset 0px -1px 0px #e0e0e0;
+  transition: 0.2s;
+}
+.form > :nth-child(odd) {
+  background: rgba(170, 206, 155, 0.2);
+  box-shadow: inset 0px -1px 0px #e0e0e0;
+  transition: 0.2s;
+}
+.secondColumn {
+  width: 45%;
+}
+.inputsProfile {
+  width: 95%;
+  border: none;
+  font-family: 'Roboto';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 15px;
+  line-height: 100%;
+  color: #373867;
+  background-color: transparent;
+  cursor: pointer;
+}
+.errorInput {
+  color: #d30404;
+  font-family: 'Roboto';
+  font-style: normal;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 21px;
+  margin: 5px 0 0 5px;
+  padding: 0%;
+  border: none;
+}
+.redButton {
+  background-color: red;
+  width: 30%;
+  margin: 15px 10%;
+  color: #373867;
+  height: 30px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+.greenButton {
+  width: 30%;
+  margin: 15px 10%;
+  height: 30px;
+  background-color: #aace9b;
+  color: #373867;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+.buttonSubmit {
+  cursor: pointer;
+  height: 30px;
+  width: 90%;
+  margin: 15px auto;
+  background-color: #aace9b !important;
+  color: #373867;
+  border: solid 1px #e0e0e0;
+  box-shadow: inset 0px -1px 0px #e0e0e0;
+  border-radius: 5px;
+}
+/* .goHome {
+  width: 30%;
+  margin: 15px 10%;
+  height: 30px;
+  color: #FFF !important;
+  background: rgb(178, 172, 172);
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+} */
+.redButton:hover,
+.greenButton:hover,
+.buttonSubmit:hover {
+  box-shadow: 0 2px 8px rgb(111, 112, 113);
+  transition: 0.2s ease;
+}
+
+.goHome {
+  width: 30%;
+  height: 30px;
+  margin: 15px 10%;
+  background: rgb(171, 170, 170);
+  color: #fff !important;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  position: relative;
+  z-index: 1;
+  overflow: hidden;
+  display: inline-block;
+}
+.goHome::after {
+  content: '';
+  background: #181823;
+  position: absolute;
+  z-index: -1;
+  padding: 16px 20px;
+  display: block;
+  top: 0;
+  bottom: 0;
+  left: -100%;
+  right: 100%;
+  -webkit-transition: all 0.35s;
+  transition: all 0.35s;
+}
+.goHome:hover::after {
+  left: 0;
+  right: 0;
+  top: 0;
+  bottom: 0;
+  -webkit-transition: all 0.35s;
+  transition: all 0.35s;
+}
