+++ src/Components/Shared/Table/table.module.css
@@ -1,5 +1,5 @@
 .container{
-    margin: 5px 5px;
+    margin: 5px auto;
     font-weight: 10px;
     font-size: 13px;
     width: 90%;
