+++ src/Components/SuperAdmins/super-admins.module.css
@@ -17,9 +17,6 @@ h2{
 div > .buttonsDeleteModal{
   margin: 10px auto;
 }
-Form > div{
-  margin: 10px 0;
-}
 .inputsForm{
   margin: 5px 0;
 }
